fix(maxwell): 修真机闭环三处崩溃（探头路径越界 + 两处 PyQt5.15 float→QSpinBox）
- recording_maxwell._streamer_binary_path：路径上溯层级 off-by-one（4→3），
  之前算成 <repo>父目录/cpp/... → connect 报「二进制不存在」→ start_session 失败
  → self.recording/stimulation=None → 闭环 update_run 全程 NoneType。这是 loop
  跑不起来的根因（探头其实已编译，路径找错了，不用重新 make）
- SignalWidget：spb_ISI/frequency/unit_time 全是 QSpinBox，但 ISI=1000/freq 等
  是 float，PyQt5.15 严格拒绝 float → 图1「Add Stimulating Signal」崩。7 处 int() 转型
- SelectStimulatingPara.show_signal / StimulateSetting.repeat_changed：
  spb_total_time.setValue(x_global/1000) 同类 float→QSpinBox 崩（图2 选参数 / 图1 重算时间）

均为 MCS 继承代码在 PyQt5.15 下的既有 bug，被 Maxwell 路径首次暴露。
验证：py_compile + 离屏构造 SignalWidget(默认/float para) + 图2 show_signal(MAXWELL) 通过；
探头路径修正后指向 <repo>/cpp/maxwell_streamer/build/maxwell_streamer。

Yundid_desktop | 2026-06-06

+++ src/SelectStimulatingPara.py
@@ -217,7 +217,7 @@ class SelectStimulatingPara(QDialog, Ui_SelectStimulatingDialog):
                 # x_global = x_global + ISI*1000
 
             self.x_aix.setRange(0, x_global)
-            self.spb_total_time.setValue(x_global/1000)
+            self.spb_total_time.setValue(int(x_global/1000))
         elif self.sys_device == SYSTEM_DEVICE.INTAN:
             self.frame_signal.hide()
             self.spb_total_time.setValue(sti_para["total_time"])
